Move stuff out of Api
