+++ src/ProductConstructionService/ProductConstructionService.Api/InitializationBackgroundService.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.DotNet.DarcLib;
 using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
-using ProductConstructionService.Api.Queue;
+using ProductConstructionService.WorkItems;
 
 namespace ProductConstructionService.Api;
