+++ src/ProductConstructionService/ProductConstructionService.WorkItems/WorkItemProcessorState.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace ProductConstructionService.Api.Queue;
+namespace ProductConstructionService.WorkItems;
 
 public enum WorkItemProcessorState
 {
