+++ src/ProductConstructionService/ProductConstructionService.WorkItems/WorkItemScopeManager.cs
@@ -1,7 +1,10 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace ProductConstructionService.Api.Queue;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+
+namespace ProductConstructionService.WorkItems;
 
 public class WorkItemScopeManager
 {
